09/11/2019

+++ src/components/SessionItem/SessionItem.js
@@ -1,11 +1,25 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
+import Swal from 'sweetalert2';
 import * as actions from '../../actions/index';
 class SessionItem extends Component {
     onClick = (e) => {
-        e.preventDefault();;
-        this.props.onDeleteSession(this.props.session.id);
+        e.preventDefault();
+        Swal.fire({
+            title: 'Bạn có chắc không?',
+            text: "Bạn không thể phục hồi lại!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Xóa phiên hỏi đáp này!',
+            cancelButtonText: 'Thoát'
+        }).then((result) => {
+            if (result.value) {
+                this.props.onDeleteSession(this.props.session.id);
+            }
+        })
     }
     render() {
         let { session } = this.props;
@@ -28,7 +42,7 @@ class SessionItem extends Component {
                             className="session-master"
                         >Chủ tọa: {session.nameOfOwner}
                         </Link>
-                        <p className="session-number">{session.questions.length === 0 ? "Chưa có câu hỏi" : `Số câu hỏi: ${session.questions.length}` }  <i className="fa fa-clock-o" aria-hidden="true"></i> {(new Date(session.createdAt)).toLocaleString()} <a href="/" onClick={this.onClick}>{session.permission ? "Xóa" : ""}</a></p>
+                        <p className="session-number">{session.questions.length === 0 ? "Chưa có câu hỏi" : `Số câu hỏi: ${session.questions.length}`}  <i className="fa fa-clock-o" aria-hidden="true"></i> {(new Date(session.createdAt)).toLocaleString()} <a href="/" onClick={this.onClick}>{session.permission ? "Xóa" : ""}</a></p>
                     </div>
                 </div>
             </React.Fragment>
@@ -38,7 +52,7 @@ class SessionItem extends Component {
 }
 const mapDispatchToProps = (dispatch, props) => {
     return {
-        onDeleteSession : (id) => {
+        onDeleteSession: (id) => {
             dispatch(actions.actDeleteSessionRequest(id));
         }
     }
